+++ NEWS.md
@@ -3,6 +3,7 @@
 ### Bug
 
 * Fix the incorrect layout of `m_button()` in the rmarkdown ([#31](https://github.com/swsoyee/r3dmol/pull/31)).
+* Fix `colorScheme="default"` overriding user-specified colors ([#34](https://github.com/swsoyee/r3dmol/pull/34))
 
 # r3dmol 0.1.2